DRY up code
