+++ src/components/Search/hooks/useFilterFeedValue.tsx
@@ -1,25 +1,7 @@
-import type {OnyxEntry} from 'react-native-onyx';
-import useFeedKeysWithAssignedCards from '@hooks/useFeedKeysWithAssignedCards';
-import useLocalize from '@hooks/useLocalize';
-import useOnyx from '@hooks/useOnyx';
-import {getFeedOptions} from '@libs/SearchUIUtils';
-import ONYXKEYS from '@src/ONYXKEYS';
-import type {SearchAdvancedFiltersForm} from '@src/types/form';
-
-function filterFeedSelector(searchAdvancedFiltersForm: OnyxEntry<SearchAdvancedFiltersForm>) {
-    return searchAdvancedFiltersForm?.feed;
-}
+import useFilterFeedData from './useFilterFeedData';
 
 function useFilterFeedValue(): string {
-    const {translate, localeCompare} = useLocalize();
-    const [allFeeds] = useOnyx(ONYXKEYS.COLLECTION.SHARED_NVP_PRIVATE_DOMAIN_MEMBER);
-    const [personalAndWorkspaceCards] = useOnyx(ONYXKEYS.DERIVED.PERSONAL_AND_WORKSPACE_CARD_LIST);
-    const [feed] = useOnyx(ONYXKEYS.FORMS.SEARCH_ADVANCED_FILTERS_FORM, {selector: filterFeedSelector});
-
-    const feedKeysWithCards = useFeedKeysWithAssignedCards();
-
-    const feedOptions = getFeedOptions(allFeeds, personalAndWorkspaceCards, translate, localeCompare, feedKeysWithCards);
-    const feedValue = feed ? feedOptions.filter((option) => feed.includes(option.value)) : [];
+    const {feedValue} = useFilterFeedData();
     return feedValue.map((item) => item.text).join(', ');
 }
